requirements.txt
flask
flask_restful
flask_cors
python_dotenv
flask_jwt_extended
pymongo
pydantic
pydantic[email]
boto3
boto3-stubs
uuid
gunicorn
flask-swagger